Attempt to get server side rendering working
Changes I had to make:
- Load in the node codemirror runmode.
This makes the syntax-highlighter tests fail because the runmode only
works in node. Maybe we can do a conditional require based on whether it's client/server

- Run all of the source files through babel to convert jsx -> js so they can be required
server side in node

- Still have a full dist build which has to be loaded in ReadMe's front end js
We shouldn't have to do this, but due to there being a .babelrc file in this
repo, our webpack config in readme attempts to load in using that. We're stuck
on an old webpack version, so the new babelrc with the old babel-loader
causes errors

- Only load `react-json-view` in componentDidMount, which only gets run client side
This is because there's a hard dependency on `window` from within this module.
This is kinda hacky but it works

- Remove React Waypoint, so all of the endpoints get rendered server side
Maybe there's a way to conditionally use Waypoint depending on whether the
initial render is client side or server side. That way we could get the benefits
of lazy rendering if it's not been initially rendered server side.

If we're to go down this road, i'd like to add `dist/` to the gitignore and
auto run the build step prepublish so we dont have to manually commit the dist files